backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
certifi
apscheduler
yfinance
ta
google-genai
pydantic
